refactor(core): Provide NoopNgZone in root factory of NgZone
This updates NgZone to be injectable in root and use NoopNgZone by
defualt. This eliminates the need for the zoneless providers in every
environment, greatly simplifying the code and eliminating failure paths
where these providers could be forgotten.
