+++ packages/core/src/render3/after_render/manager.ts
@@ -7,10 +7,8 @@
  */
 
 import {TracingAction, TracingService, TracingSnapshot} from '../../application/tracing';
-import {
-  ChangeDetectionScheduler,
-  NotificationSource,
-} from '../../change_detection/scheduling/zoneless_scheduling';
+import {NotificationSource} from '../../change_detection/scheduling/tokens';
+import {ChangeDetectionScheduler} from '../../change_detection/scheduling/change_detection_scheduler';
 import {inject} from '../../di/injector_compatibility';
 import {ɵɵdefineInjectable} from '../../di/interface/defs';
 import {ErrorHandler} from '../../error_handler';
