+++ packages/core/test/bundling/animations-standalone/bundle.golden_symbols.json
@@ -78,7 +78,7 @@
       "DISABLE_ANIMATIONS_FLAG",
       "DI_DECORATOR_FLAG",
       "DOCUMENT",
-      "DOCUMENT2",
+      "DOCUMENT",
       "DefaultDomRenderer2",
       "DestroyRef",
       "DomAdapter",
@@ -306,7 +306,6 @@
       "_BrowserDomAdapter",
       "_CACHED_BODY",
       "_ChainedInjector",
-      "_ChangeDetectionScheduler",
       "_ComponentFactory",
       "_ComponentFactoryResolver",
       "_ComponentRef",
@@ -324,7 +323,6 @@
       "_IS_WEBKIT",
       "_InjectionToken",
       "_Injector",
-      "_NgZone",
       "_NodeInjector",
       "_NodeInjectorDestroyRef",
       "_NodeInjectorFactory",
@@ -528,7 +526,6 @@
       "eraseStyles",
       "errorContext",
       "errorHandler",
-      "errorHandlerEnvironmentInitializer",
       "errorNotification",
       "execFinalizer",
       "executeCheckHooks",
@@ -816,7 +813,6 @@
       "projectNodes",
       "provideAnimations",
       "provideZonelessChangeDetection",
-      "provideZonelessChangeDetectionInternal",
       "providerToFactory",
       "providerToRecord",
       "publishSignalConfiguration",
