+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -70,7 +70,7 @@
       "DISABLED",
       "DI_DECORATOR_FLAG",
       "DOCUMENT",
-      "DOCUMENT2",
+      "DOCUMENT",
       "DefaultDomRenderer2",
       "DefaultIterableDiffer",
       "DefaultIterableDifferFactory",
@@ -303,7 +303,6 @@
       "_BrowserDomAdapter",
       "_BrowserGetTestability",
       "_ChainedInjector",
-      "_ChangeDetectionScheduler",
       "_ComponentFactory",
       "_ComponentFactoryResolver",
       "_ComponentRef",
@@ -337,7 +336,6 @@
       "_NgForOfContext",
       "_NgModuleFactory",
       "_NgModuleRef",
-      "_NgZone",
       "_NodeInjector",
       "_NodeInjectorDestroyRef",
       "_NodeInjectorFactory",
@@ -580,7 +578,6 @@
       "epoch",
       "errorContext",
       "errorHandler",
-      "errorHandlerEnvironmentInitializer",
       "errorNotification",
       "execFinalizer",
       "executeCheckHooks",
@@ -938,7 +935,6 @@
       "profiler",
       "profilerCallbacks",
       "projectNodes",
-      "provideZonelessChangeDetectionInternal",
       "providerToFactory",
       "providerToRecord",
       "providersResolver",
