+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -71,7 +71,7 @@
       "DISABLED",
       "DI_DECORATOR_FLAG",
       "DOCUMENT",
-      "DOCUMENT2",
+      "DOCUMENT",
       "DefaultDomRenderer2",
       "DefaultIterableDiffer",
       "DefaultIterableDifferFactory",
@@ -307,7 +307,6 @@
       "_BrowserDomAdapter",
       "_BrowserGetTestability",
       "_ChainedInjector",
-      "_ChangeDetectionScheduler",
       "_ComponentFactory",
       "_ComponentFactoryResolver",
       "_ComponentRef",
@@ -339,7 +338,6 @@
       "_NgForOfContext",
       "_NgModuleFactory",
       "_NgModuleRef",
-      "_NgZone",
       "_NodeInjector",
       "_NodeInjectorDestroyRef",
       "_NodeInjectorFactory",
@@ -579,7 +577,6 @@
       "epoch",
       "errorContext",
       "errorHandler",
-      "errorHandlerEnvironmentInitializer",
       "errorNotification",
       "execFinalizer",
       "executeCheckHooks",
@@ -933,7 +930,6 @@
       "profiler",
       "profilerCallbacks",
       "projectNodes",
-      "provideZonelessChangeDetectionInternal",
       "providerToFactory",
       "providerToRecord",
       "providersResolver",
