+++ packages/core/test/bundling/hydration/bundle.golden_symbols.json
@@ -59,7 +59,7 @@
       "DISCONNECTED_NODES",
       "DI_DECORATOR_FLAG",
       "DOCUMENT",
-      "DOCUMENT2",
+      "DOCUMENT",
       "DefaultDomRenderer2",
       "DestroyRef",
       "DomAdapter",
@@ -247,7 +247,6 @@
       "_BehaviorSubject",
       "_BrowserDomAdapter",
       "_ChainedInjector",
-      "_ChangeDetectionScheduler",
       "_ComponentFactory",
       "_ComponentFactoryResolver",
       "_ComponentRef",
@@ -267,7 +266,6 @@
       "_HttpResponseBase",
       "_InjectionToken",
       "_Injector",
-      "_NgZone",
       "_NodeInjector",
       "_NodeInjectorDestroyRef",
       "_NodeInjectorFactory",
@@ -464,7 +462,6 @@
       "epoch",
       "errorContext",
       "errorHandler",
-      "errorHandlerEnvironmentInitializer",
       "errorNotification",
       "escapeCommentText",
       "execFinalizer",
@@ -631,7 +628,7 @@
       "isInI18nBlock",
       "isInInjectionContext",
       "isInSkipHydrationBlock",
-      "isInSkipHydrationBlock2",
+      "isInSkipHydrationBlock",
       "isInlineTemplate",
       "isInputBinding",
       "isInteropObservable",
@@ -730,7 +727,6 @@
       "profilerCallbacks",
       "projectNodes",
       "provideClientHydration",
-      "provideZonelessChangeDetectionInternal",
       "providerToFactory",
       "providerToRecord",
       "publishSignalConfiguration",
