+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -71,7 +71,7 @@
       "DEFAULT_SERIALIZER",
       "DI_DECORATOR_FLAG",
       "DOCUMENT",
-      "DOCUMENT2",
+      "DOCUMENT",
       "DefaultDomRenderer2",
       "DefaultRouteReuseStrategy",
       "DefaultTitleStrategy",
@@ -361,7 +361,6 @@
       "_CanActivate",
       "_CanDeactivate",
       "_ChainedInjector",
-      "_ChangeDetectionScheduler",
       "_ChildActivationEnd",
       "_ChildActivationStart",
       "_ComponentFactory",
@@ -398,7 +397,6 @@
       "_NavigationStart",
       "_NgModuleFactory",
       "_NgModuleRef",
-      "_NgZone",
       "_NodeInjector",
       "_NodeInjectorDestroyRef",
       "_NodeInjectorFactory",
@@ -693,7 +691,6 @@
       "equalSegments",
       "errorContext",
       "errorHandler",
-      "errorHandlerEnvironmentInitializer",
       "errorNotification",
       "exactMatchOptions",
       "execFinalizer",
@@ -1077,7 +1074,6 @@
       "projectNodes",
       "provideRouter",
       "provideZonelessChangeDetection",
-      "provideZonelessChangeDetectionInternal",
       "providerToFactory",
       "providerToRecord",
       "publishSignalConfiguration",
