+++ packages/core/test/bundling/standalone_bootstrap/bundle.golden_symbols.json
@@ -48,7 +48,7 @@
       "DEFAULT_LOCALE_ID",
       "DI_DECORATOR_FLAG",
       "DOCUMENT",
-      "DOCUMENT2",
+      "DOCUMENT",
       "DefaultDomRenderer2",
       "DestroyRef",
       "DomAdapter",
@@ -205,7 +205,6 @@
       "_BehaviorSubject",
       "_BrowserDomAdapter",
       "_ChainedInjector",
-      "_ChangeDetectionScheduler",
       "_ComponentFactory",
       "_ComponentFactoryResolver",
       "_ComponentRef",
@@ -222,7 +221,6 @@
       "_HelloWorld",
       "_InjectionToken",
       "_Injector",
-      "_NgZone",
       "_NodeInjector",
       "_NodeInjectorDestroyRef",
       "_NodeInjectorFactory",
@@ -380,7 +378,6 @@
       "epoch",
       "errorContext",
       "errorHandler",
-      "errorHandlerEnvironmentInitializer",
       "errorNotification",
       "execFinalizer",
       "executeCheckHooks",
@@ -586,7 +583,6 @@
       "profilerCallbacks",
       "projectNodes",
       "provideZonelessChangeDetection",
-      "provideZonelessChangeDetectionInternal",
       "providerToFactory",
       "providerToRecord",
       "publishSignalConfiguration",
